added new module contains DataStore class and gatherd storing data structures to it.
